requirements.txt
# FastAPI and server
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.10.3
pydantic-settings==2.6.1

# Machine Learning
scikit-learn==1.5.2
joblib==1.5.1
numpy==2.1.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10  # PostgreSQL
pymysql==1.1.1          # MySQL
mysqlclient==2.2.4      # Alternative MySQL driver
alembic==1.14.0

# Caching
cachetools==5.5.0

# Monitoring and utilities
psutil==6.1.0
python-multipart==0.0.18

# Development and testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Optional dependencies
redis==5.2.1  # For distributed caching
prometheus-client==0.21.1  # For metrics
structlog==24.4.0  # For structured logging
